phase1/requirements.txt
# Flask and Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5

# Payment Processing
stripe==7.11.0

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9

# Environment Configuration
python-dotenv==1.0.0

# Data Validation
pydantic==2.5.3

# HTTP Client
requests==2.31.0

# Date/Time
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0

# Async Support
aiohttp==3.13.3
asyncio==3.4.3
